src/main/java/edu/buaa/park/ParkingBoy.java
package edu.buaa.park;

import java.util.List;

public class ParkingBoy implements ParkingWorker {
    protected List<ParkPlace> parkPlaces;
    private final ParkingLotChooser parkingLotChooser;

    public ParkingBoy(List<ParkPlace> parkPlaces, ParkingLotChooser parkingLotChooser) {
        this.parkPlaces=parkPlaces;
        this.parkingLotChooser = parkingLotChooser;
    }
    public Ticket park(Car car) {
      return parkingLotChooser.getAvailablePark(parkPlaces).parkCar(car);
    }

    public Integer getAvailableNum() {
        int availableNum=0;
        for(ParkPlace parkPlace:parkPlaces){
            availableNum+=parkPlace.getAvailableNum();
        }
       return availableNum;
    }
    public Car fetch(Ticket ticket) {
        Car fetchedCar=null;
        for(ParkPlace parkPlace:parkPlaces){
            fetchedCar=parkPlace.fecthCar(ticket);
            if(fetchedCar!=null){return fetchedCar;}
        }
        throw new NoCarException("没有此车");
    }

    public List<ParkPlace> getParkPlaces(){
        return parkPlaces ;
    }

}
